+++ test/test.binding.dom.js
@@ -22,14 +22,17 @@ describe('binding to dom element manually', function() {
 
     o_O.bindElementToRule(el, "css", "{color: color()}", o)
 
-    expect($("#el").css("color")).to.be('rgb(3, 3, 3)')
+    o_O.nextFrame(function() {
+      expect($("#el").css("color")).to.be('rgb(3, 3, 3)')
 
-    o.color("rgb(3, 3, 4)")
-    
-    setTimeout(function() {
-      expect($("#el").css("color")).to.be("rgb(3, 3, 4)")
-      done()
-    }, 0)
+      o.color("rgb(3, 3, 4)")
+
+       o_O.nextFrame(function() {
+        expect($("#el").css("color")).to.be("rgb(3, 3, 4)")
+        done()
+      })
+    })
+
 
   })
 
@@ -77,19 +80,22 @@ describe('binding to dom element manually', function() {
 
   
   describe('custom bindings', function() {    
-    it('work for setting', function() {
+    it('work for setting', function(done) {
       
       o_O.bindings.color = function(val, $el) {
         $el.css("color", val)
-
       }
       
       o_O.bindElementToRule(el, "color", "'red'")
       
-      expect($(el).css("color")).to.be("rgb(255, 0, 0)")
+      o_O.nextFrame(function() {
+        expect($(el).css("color")).to.be("rgb(255, 0, 0)")
+        done()
+      })
+      
     })
     
-    it('works with properties', function() {
+    it('works with properties', function(done) {
       
       o_O.bindings.color = function(val, $el) {
         $el.css("color", val)
@@ -97,28 +103,38 @@ describe('binding to dom element manually', function() {
       
       var o = { red: o_O('red') }
       o_O.bindElementToRule(el, "color", "red()", o)
-      expect($(el).css("color")).to.be("rgb(255, 0, 0)")
+      o_O.nextFrame(function() {
+        expect($(el).css("color")).to.be("rgb(255, 0, 0)")
+        done()
+      })
     })
     
     
-    it('works for evented custom bindings', function() {
-      var o = { check: function(e) { 
-        last = e 
-      } }
-            
-      o_O.bindings.wham = function(callback, $el) {
-        callback('woah')
-        expect(this).to.be(o)
-        expect($el[0]).to.be(el)
-      }
-      
-
-      o_O.bindElementToRule(el, "wham", "check", o)
-      
-      expect(last).to.be('woah')
-      
-    })
-    
+    // it('works for evented custom bindings', function(done) {
+    //   var last 
+    //   var o = { 
+    //     check: function(e) { 
+    //       last = e 
+    //     } 
+    //   }
+    //         
+    //   o_O.bindings.wham = function(callback, $el) {
+    //     callback('woah')
+    //     expect(this).to.be(o)
+    //     expect($el[0]).to.be(el)
+    //   }
+    //   
+    //   
+    //   o_O.bindElementToRule(el, "wham", "check", o)
+    //   
+    //   o_O.nextFrame(function() {
+    //     expect(last).to.be('woah')
+    //     done()
+    //   })
+    //   
+    //   
+    // })
+    // 
     
   })    
       
@@ -141,32 +157,33 @@ describe('call binding', function() {
     __count = 0
   })
   
-  it('gets called once when set as a function call', function(done) {
+  it('gets called once when set as a function call', function() {
     $(el).attr('data-bind', 'call: __incr')
     expect(__count).to.eql(0)
     o_O.bind({}, el)
-    setTimeout(function() {
+    // o_O.nextFrame(function() {
       expect(__count).to.eql(1)
       __dependency.incr()
-      setTimeout(function() {
+      // o_O.nextFrame(function() {
         expect(__count).to.eql(1)
-        done()
-      }, 0)
-    }, 0)
+        // done()
+    //   })
+    // })
   })
 
-  it('gets called twice when set as a called function ', function(done) {
+  it('gets called twice when set as a called function ', function() {
     $(el).attr('data-bind', 'call: __incr()')
     expect(__count).to.eql(0)
     o_O.bind({}, el)
-    setTimeout(function() {
+    // o_O.nextFrame(function() {
       expect(__count).to.eql(1)
       __dependency.incr()
-      setTimeout(function() {
-        expect(__count).to.eql(2)
-        done()
-      }, 0)
-    }, 0)
+      // done()
+      // o_O.nextFrame(function() {
+      //   expect(__count).to.eql(2)
+      //   done()
+      // })
+    
   })
   
 })